nmp: pf: change int to u8 in mng_cmd_params
- reset parameter was defined as int.
- changed to u8.
- no functioal change involved in this patch

Reviewed-on: http://vgitil04.il.marvell.com:8080/52317
Tested-by: Liron Himi <lironh@marvell.com>

+++ src/mng/lf/mng_cmd_desc.h
@@ -153,7 +153,7 @@ struct mgmt_cmd_params {
 		} pf_ingress_data_q_add;
 
 		struct {
-			int	reset;
+			u8	reset;
 		} pf_get_statistics;
 
 		struct {
